requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain_ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain-google-genai
langchain_groq
fastapi
langserve
sse_starlette